Move away from radix-based implementation
